switches encryption to AES (#9)
To clearly define what is need on the client side, a small javascript example was created to show how the encryption works and what the encryption string should look like. Additionally, more validation and testing has been added. The schema of the encryption may change based on feedback from others.